Add platform definition for debian-9-armhf
This commit adds a plaform definition based on debian-8-armhf and
debian-8-amd64. In addition to setting up for cross compilation, the
image is also configured to build inside of Docker containers.

The source for build tools is currently hardcoded to a repository that I've
created on bintray.com:

  https://bintray.com/sharpie/pl-build-tools-deb

This is not ideal, but 100000000000% better than an internal repo that noone
outside of the VPN can access.